Create coretag-c:remove
It is used for removing the specified variable from a particular scope. This action is not particularly helpful, but it can be used for ensuring that a JSP can also clean up any scope resources.

The <c:remove > tag removes the variable from either a first scope or a specified scope.