Create fn:trim( )
The fn:trim() function removes the blank spaces from both the ends of a string. It mainly used for ignoring the blank spaces from both the ends of string.The fn:trim() function removes the blank spaces from both the ends of a string. It mainly used for ignoring the blank spaces from both the ends of string.